+++ src/net/session.cc
@@ -36,6 +36,12 @@ using namespace residue;
 const std::string Session::PACKET_DELIMITER = "\r\n\r\n";
 const std::size_t Session::PACKET_DELIMITER_SIZE = Session::PACKET_DELIMITER.size();
 
+const std::unordered_map<Response::StatusCode, std::string> Session::STANDARD_RESPONSES = {
+    { Response::StatusCode::STATUS_OK, "{r:0}" },
+    { Response::StatusCode::CONTINUE, "{r:0}" },
+    { Response::StatusCode::BAD_REQUEST, "{r:1}" },
+};
+
 Session::Session(tcp::socket&& socket,
                  RequestHandler* requestHandler) :
     m_socket(std::move(socket)),
@@ -145,14 +151,8 @@ void Session::write(const char* data,
 
 void Session::writeStatusCode(const Response::StatusCode& r)
 {
-    std::stringstream ss;
-    if (r == Response::StatusCode::STATUS_OK || r == Response::StatusCode::CONTINUE) {
-        ss << "{r:" << 0 << "}";
-    } else {
-        ss << "{r:" << 1 << ",c:" << static_cast<unsigned short>(r) << "}";
-    }
-    ss << Session::PACKET_DELIMITER;
-    write(ss.str().c_str(), ss.str().length());
+    std::string resp(Session::STANDARD_RESPONSES.at(r) + Session::PACKET_DELIMITER);
+    write(resp.c_str(), resp.length());
 }
 
 void Session::write(const std::string& s)